Moving pigeon-hole test cases to own class.
